Moved HTML to JS
We moved a large block of HTML to JS so that we can inject the info several times with new info. We created a new function called forecast that will be called from the bottom. Made a row var to create a grid. Made a for each loop to go through the array of our days. Each time the loop runs, it puts the date name on the weather to override the forecast HTML and concatenate the HTML. Uses the same HTML. Involves making the code smarter. Have to use a closing div at the end of our function.